feat(catalog): decimal line-item quantities + admin cleanup of priceless services
Allow fractional invoice line-item quantities so services billed by area or
length (e.g. 1.5 m² of tarpaulin) can be entered directly instead of being
rejected as invalid — which previously blocked printing the document even
though the computed total was correct.

- Change InvoiceLineItem.Quantity from int to float64 across the OpenAPI
  contract, Go domain, cost/margin calc, PDF report, and web zod validation.
- Format fractional quantities with a Serbian comma separator (1,5) in both
  the backend PDF and the web print sheet; whole numbers stay decimal-free.
- Add step="any" to the quantity input so the browser accepts decimals.

Also add an admin-only catalog cleanup action that deletes service items left
without both an input (purchase) and output (sale) price:

- New POST /catalog-items/cleanup endpoint (admin) + store methods.
- "Clean up empty services" button on the catalog page with a confirm dialog
  and result toast (sr/en strings).

Tests: fractional-quantity profit + PDF row formatting, catalog cleanup at the
store and API layers (incl. admin gating).

+++ apps/web/src/lib/work-orders/validation.ts
@@ -86,7 +86,9 @@ const invoiceLineItemSchema = z
     id: z.string(),
     kind: invoiceLineItemKindEnum,
     description: z.string(),
-    quantity: z.number().int().positive(),
+    // Fractional quantities are allowed so services billed by area or length
+    // (e.g. 1.5 m² of tarpaulin) can be entered directly, not just whole pieces.
+    quantity: z.number().positive(),
     unit: invoiceUnitEnum,
     unitPrice: z.number().min(0),
     // Per-unit cost. Server-derived for catalog lines; for ad-hoc lines an admin
